+++ .github/workflows/dataprep.yml
@@ -1,0 +1,40 @@
+name: Run Data prep
+
+on:
+  push:
+    branches:
+      - main
+      - dev
+  pull_request:
+  release:
+
+jobs:
+  build:
+    runs-on: ${{ matrix.os }}
+    defaults:
+      run:
+        shell: bash -l {0}
+    strategy:
+      matrix:
+        os: ['ubuntu-latest', 'windows-latest', 'macos-latest']
+
+    steps:
+      - uses: actions/checkout@v3
+
+      - name: Install Conda environment
+        uses: mamba-org/provision-with-micromamba@main
+        with:
+          cache-downloads: true
+          cache-env: true
+          environment-file: environment.yml
+          environment-name: loopy
+
+      - name: Install ncurses if Windows
+        if: matrix.os == 'windows-latest'
+        run: |
+          pip install windows-curses
+
+      - name: Test
+        run: |
+          micromamba install -c conda-forge pytest pytest-cov
+          pytest
